[MOD] #8 Viewmodel 추가
