+++ app/src/main/java/com/solux/dorandoran/domain/entity/DiscussionPageEntity.kt
@@ -3,6 +3,7 @@ package com.solux.dorandoran.domain.entity
 data class DiscussionPageEntity (
 
     val id: Int,
+    val name: String,
     val bookTitle: String,
     val discussionTopic: String,
     val bookImageUrl: String,
